+++ vitest.config.js
@@ -16,6 +16,16 @@ export default defineConfig({
     projects: [
       {
         extends: true,
+        name: 'unit',
+        include: ['src/**/*.test.js'],
+        exclude: ['e2e/**'],
+        test: {
+          name: 'unit',
+          globals: true,
+        },
+      },
+      {
+        extends: true,
         plugins: [
           // The plugin will run tests for the stories defined in your Storybook config
           // See options at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon#storybooktest